+++ tests/test75.c
@@ -1,7 +1,8 @@
 #include <stdio.h>    /* printf */
 #include "utstring.h"
 
-int main() {
+int main()
+{
     UT_string *s,*t;
     char V_TestStr[] = "There are two needle\0s in this \0haystack with needle\0s.";
     char V_NeedleStr[] = "needle\0s";
@@ -13,21 +14,19 @@ int main() {
     utstring_new(t);
 
     utstring_bincpy(s, V_TestStr, sizeof(V_TestStr)-1);
-    printf("\"%s\" len=%u\n", utstring_body(s), utstring_len(s));
+    printf("\"%s\" len=%u\n", utstring_body(s), (unsigned)utstring_len(s));
     utstring_bincpy(t, V_NeedleStr, sizeof(V_NeedleStr)-1);
-    printf("\"%s\" len=%u\n", utstring_body(t), utstring_len(t));
+    printf("\"%s\" len=%u\n", utstring_body(t), (unsigned)utstring_len(t));
 
     V_FindCnt = 0;
     V_FindPos = -1;
-    do
-    {
-        V_FindPos = utstring_findR(s, 
-                                   V_FindPos, 
-                                   utstring_body(t), 
+    do {
+        V_FindPos = utstring_findR(s,
+                                   V_FindPos,
+                                   utstring_body(t),
                                    utstring_len(t));
         printf("utstring_findR()=%ld\n", V_FindPos);
-        if (V_FindPos >= 0)
-        {
+        if (V_FindPos >= 0) {
             V_FindPos--;
             V_FindCnt++;
         }
